requirements.txt
streamlit==1.44.0
google-genai>=1.0.0
PyPDF2>=3.0.1
python-dotenv>=0.21.0
chromadb>=0.5.0
pydantic>=2.0.0,<3.0.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
